Finished experience level math
I tried to make a compromise between the gen I and II method of
computing the exp for the next level on evolution, and a pure lookup
table with no math. What I have done is have the algorithms populate
lookup tables at runtime. This allows us to change the algorithms if we
want, without having to go through and change literally all 100 entries
in a lookup table.