Overhaul of Master branch
After all of the work we did earlier this week with the mergers, the
master branch needed a lot of cleaning up. I'll try and include a change
log of everything that is different about the project now.

####Major
+ IConsole.cs added
+ IConsole is like an improved console, I'll try and add methods to it
to help us in using more complicated formatting. Like color and text
alignment.
+ objects data now exists in the model
+ Types.xml is now read into the game
+ Error catching. Try-Catch statements now ensure that the program has
data that it needs for the game to start. There is a boolean check then
performed, and the game will start on a passed check.

####Minor
+ Corrected namespace issue in Types.cs
+ Types is now public
+ Added necessary constructors for bringing in Type data
+ Added Types array to the Model
+ All game math function calls are now wrapped in the gameLoadPopulate()
method.