Types now fully integer based
Pokemon types are now fully supporting being integer based. Please note
going forward that if you are interacting with types in the future, that
"-1" represents a null type. previously no text represented this. But
given types are 0 based, it was important that there be some way to
nullify the unused indexes within the types array. If you are going
through the types array, ensure that you are checking that the current
index is not -1.