Began adding code for building command
I hit a wall when coding this functionality. In order to get the left
side of the string, I would need to know how many characters exist
before a potential operand is entered, so that it can be used as the
case. I found this was very hard to accomplish in C#. I'm going to try
something by creating a C++ library.