Update the .cs files
In some cases I just made changes so that Git will register them as a
"new" commit so that they can be brought into the master branch. However
I also took it as a moment to clear out a lot of unnecessary resources
that were being used by a lot of classes (like generic collections or
LINQ). I'll have to do a check to ensure that everything is building
correctly once I get them into visual studio and try and do a build.