brought in the mongoDB drivers
I have changed my mind from where I was previously on this issue, and
have decided to go with MongoDB as our database. There are a few reasons
for this, which I will extrapolate on in the wiki.

I can't be sure this will actually bring in the whole MongoDB driver for
everyone however. If you don't actually have a reference to MongoDB.Bson
and MongoDB.Driver in your Pokemon project references, use the NuGet
package manager to install the official MongoDB driver for C#. If you
have an issues with this just let me know.