Working now
I'm at a loss. I've done as far as I can tell the same thing I was doing
last night and it's working without a hitch. I'm assuming I was spelling
something wrong or not setting up my collection correctly. In either
event I have something to refer to now for when I'm running queries in
the future.