Simple Mongo Query
This adds in a simple mongo query.  At this point the database just
contains the types data. This runs a string builder,  then converts it
to a string and puts it into the body of a message box. This is
essentially just a sanity check, to show that we have data coming out of
the database.