additional game math functions
I was thinking about this, and the way I originally planned out the
functions was to pass them an array and fill them with everything.
However, I wasn't sure if that is really how I wanted to do it the more
I thought about it, so I made some new functions that take in an integer
and return the value of that single level.