Region of origin looping is complete
The list of images now loops through properly, and the user can go
backward and forward with both of the buttons. I may want to refactor
how this works though. The two buttons do incredibly similar things, so
they would make sense as one button. secondly they both set the name and
description field, which could be a method of it's own.