clean up and professor talk
I cleaned up some of the code behind the Begin Adventure from as well as
placed things into regions to help keep better track of what was what.
Also, a panel was added to the form which will be set to visible when
the player has selected their starting info.