ready for text and images
I have most of the visual effects I need for the professor talk in
place. Up next I need to dig into the creation and presentation of the
pokemon to user, as I'm know I'm missing a few needed features for them,
specifically the creation of unique ids in the Database for the pokemon
(critical) as well as all of the graphical images (which I have some of
started back on my desktop).