successful addition basepokemon collection
There is now a collection of base pokemon in the database. Everything
was added up into the database succesfully. One little catch is that
issue I ran into with types forever ago. I'll need to make sure that
monotypes have a secondary "type" of -1 to denote that it doesn't have a
second type, because otherwise every monotype will actually have a
secondary type of "Normal"